+++ changelog.md
@@ -4,6 +4,7 @@
  - Emitter now has `.error` method for emitting errors
  - Emitter object also now has `.error` method
  - Properties now might have a current error along with current value
+ - New method `Kefir.constantError`
 
 ## 0.4.2